+++ Sources/XYZ.cpp
@@ -50,7 +50,7 @@ int lire_XYZ_Para(int b_z, int rang, int P, std::string source, struct Vecteur_P
       return i;
 }
 
-void ecrire_XYZ_Para_local(std::string cible, std::vector<u64> &ids, struct Vecteur_Para &pos, int n_local) {
+void ecrire_XYZ_Para_local(std::string cible, std::vector<u32> &ids, struct Vecteur_Para &pos, int n_local) {
       std::ofstream fichier;
       fichier.open(cible);
